// packages/datadog-api-client-v1/models/SLOThreshold.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SLOTimeframe } from './SLOTimeframe';
/**
* SLO thresholds (target and optionally warning) for a single time window.
*/
export declare class SLOThreshold {
    /**
    * The target value for the service level indicator within the corresponding timeframe.
    */
    'target': number;
    /**
    * A string representation of the target that indicates its precision. It uses trailing zeros to show significant decimal places (e.g. `98.00`).  Always included in service level objective responses. Ignored in create/update requests.
    */
    'targetDisplay'?: string;
    'timeframe': SLOTimeframe;
    /**
    * The warning value for the service level objective.
    */
    'warning'?: number;
    /**
    * A string representation of the warning target (see the description of the `target_display` field for details).  Included in service level objective responses if a warning target exists. Ignored in create/update requests.
    */
    'warningDisplay'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SLOThreshold;
    static serialize(data: SLOThreshold): {
        [key: string]: any;
    };
    constructor();
}